# Improve performance of computing filter masks