This speeds up computing filter masks for integer and date filters, mostly by optimizing the case where the array is known to have no null values, so the null check can be skipped.

Benchmark results before optimizations:

| Method                                  | NumRows | Mean         | Error       | StdDev     |
|---------------------------------------- |-------- |-------------:|------------:|-----------:|
| **ComputeIntEqualityMask**                  | **100**     |     **1.384 μs** |   **0.0482 μs** |  **0.0026 μs** |
| ComputeIntEqualityMaskWithNullableArray | 100     |     3.139 μs |   0.0731 μs |  0.0040 μs |
| ComputeIntRangeMask                     | 100     |     1.513 μs |   0.7365 μs |  0.0404 μs |
| ComputeIntRangeMaskWithNullableArray    | 100     |     3.298 μs |   0.5293 μs |  0.0290 μs |
| ComputeDateRangeMask                    | 100     |     2.178 μs |   0.6386 μs |  0.0350 μs |
| ComputeDateRangeMaskWithNullableArray   | 100     |     3.743 μs |   0.4000 μs |  0.0219 μs |
| ComputeStringMask                       | 100     |     6.862 μs |   0.7430 μs |  0.0407 μs |
| ComputeStringMaskWithNullableArray      | 100     |    10.479 μs |   0.5818 μs |  0.0319 μs |
| **ComputeIntEqualityMask**                  | **1000**    |    **12.969 μs** |   **0.4129 μs** |  **0.0226 μs** |
| ComputeIntEqualityMaskWithNullableArray | 1000    |    33.011 μs |  21.1007 μs |  1.1566 μs |
| ComputeIntRangeMask                     | 1000    |    15.100 μs |   2.3441 μs |  0.1285 μs |
| ComputeIntRangeMaskWithNullableArray    | 1000    |    33.589 μs |  14.4407 μs |  0.7915 μs |
| ComputeDateRangeMask                    | 1000    |    24.101 μs |   4.3073 μs |  0.2361 μs |
| ComputeDateRangeMaskWithNullableArray   | 1000    |    43.769 μs |   1.4644 μs |  0.0803 μs |
| ComputeStringMask                       | 1000    |    75.660 μs |   7.8289 μs |  0.4291 μs |
| ComputeStringMaskWithNullableArray      | 1000    |   103.037 μs |   6.3109 μs |  0.3459 μs |
| **ComputeIntEqualityMask**                  | **10000**   |   **135.234 μs** |   **5.0512 μs** |  **0.2769 μs** |
| ComputeIntEqualityMaskWithNullableArray | 10000   |   312.840 μs |   8.2627 μs |  0.4529 μs |
| ComputeIntRangeMask                     | 10000   |   157.829 μs |  16.5456 μs |  0.9069 μs |
| ComputeIntRangeMaskWithNullableArray    | 10000   |   344.848 μs | 143.3870 μs |  7.8595 μs |
| ComputeDateRangeMask                    | 10000   |   292.337 μs |  30.8814 μs |  1.6927 μs |
| ComputeDateRangeMaskWithNullableArray   | 10000   |   458.569 μs |   6.8777 μs |  0.3770 μs |
| ComputeStringMask                       | 10000   |   751.143 μs | 184.5456 μs | 10.1156 μs |
| ComputeStringMaskWithNullableArray      | 10000   | 1,022.997 μs |  41.2680 μs |  2.2620 μs |

Benchmark results after:

| Method                                  | NumRows | Mean           | Error        | StdDev      |
|---------------------------------------- |-------- |---------------:|-------------:|------------:|
| **ComputeIntEqualityMask**                  | **100**     |       **203.0 ns** |     **12.95 ns** |     **0.71 ns** |
| ComputeIntEqualityMaskWithNullableArray | 100     |     3,512.8 ns |    170.98 ns |     9.37 ns |
| ComputeIntRangeMask                     | 100     |       303.0 ns |      6.62 ns |     0.36 ns |
| ComputeIntRangeMaskWithNullableArray    | 100     |     3,157.9 ns |     96.30 ns |     5.28 ns |
| ComputeDateRangeMask                    | 100     |       235.3 ns |     28.08 ns |     1.54 ns |
| ComputeDateRangeMaskWithNullableArray   | 100     |     3,039.7 ns |  1,045.57 ns |    57.31 ns |
| ComputeStringMask                       | 100     |     8,166.5 ns |  2,223.18 ns |   121.86 ns |
| ComputeStringMaskWithNullableArray      | 100     |    10,052.3 ns |    814.14 ns |    44.63 ns |
| **ComputeIntEqualityMask**                  | **1000**    |     **1,642.2 ns** |    **431.82 ns** |    **23.67 ns** |
| ComputeIntEqualityMaskWithNullableArray | 1000    |    33,326.4 ns |  1,609.16 ns |    88.20 ns |
| ComputeIntRangeMask                     | 1000    |     2,915.8 ns |     93.81 ns |     5.14 ns |
| ComputeIntRangeMaskWithNullableArray    | 1000    |    31,064.6 ns |  5,060.87 ns |   277.40 ns |
| ComputeDateRangeMask                    | 1000    |     2,117.3 ns |     70.90 ns |     3.89 ns |
| ComputeDateRangeMaskWithNullableArray   | 1000    |    34,534.9 ns |  6,750.05 ns |   369.99 ns |
| ComputeStringMask                       | 1000    |    73,417.6 ns |  4,507.39 ns |   247.07 ns |
| ComputeStringMaskWithNullableArray      | 1000    |   100,192.4 ns |    565.15 ns |    30.98 ns |
| **ComputeIntEqualityMask**                  | **10000**   |    **18,051.6 ns** |  **2,170.90 ns** |   **118.99 ns** |
| ComputeIntEqualityMaskWithNullableArray | 10000   |   329,713.7 ns |  4,769.60 ns |   261.44 ns |
| ComputeIntRangeMask                     | 10000   |    31,168.1 ns |  2,605.88 ns |   142.84 ns |
| ComputeIntRangeMaskWithNullableArray    | 10000   |   309,563.2 ns |  9,139.37 ns |   500.96 ns |
| ComputeDateRangeMask                    | 10000   |    35,076.7 ns | 16,668.99 ns |   913.68 ns |
| ComputeDateRangeMaskWithNullableArray   | 10000   |   372,787.6 ns |  6,483.50 ns |   355.38 ns |
| ComputeStringMask                       | 10000   |   749,768.3 ns | 54,944.34 ns | 3,011.68 ns |
| ComputeStringMaskWithNullableArray      | 10000   | 1,012,841.5 ns | 29,622.43 ns | 1,623.71 ns |